It is possible in the old way, but only if the import is splitted.
