--- zuks/__init__.py
@@ -1,5 +1,0 @@
-import os
-BASE_DIR = os.path.dirname(os.path.dirname(__file__))
-SETTINGS_DIR = os.path.dirname(__file__)
-PROJECT_PATH = os.path.abspath(os.path.join(SETTINGS_DIR, os.pardir))
-
