+++ zuks/settings.py
@@ -10,9 +10,11 @@ https://docs.djangoproject.com/en/1.6/ref/settings/
 
 # Build paths inside the project like this: os.path.join(BASE_DIR, ...)
 import os
-from zuks import BASE_DIR, SETTINGS_DIR, PROJECT_PATH
 from django.conf.global_settings import TEMPLATE_CONTEXT_PROCESSORS
 
+BASE_DIR = os.path.dirname(os.path.dirname(__file__))
+SETTINGS_DIR = os.path.dirname(__file__)
+PROJECT_PATH = os.path.abspath(os.path.join(SETTINGS_DIR, os.pardir))
 # Application definition
 
 INSTALLED_APPS = (
